Fix version cmd using go linker
Initially the version command was set to "0.0.0" with the hope that
I would eventually get back to actually correctly configuring it.

Luckily, that time has come and Ive opted to reference a few projects
and articles with how to set the -X linker option. Referenced the
following resources for help:
  https://github.com/databus23/helm-diff/
  https://blog.cloudflare.com/setting-go-variables-at-compile-time/

This will now allow me to correctly reference the version via git
tags as well as any versions that are cut outside the normal release
cycle for testing by appending the git SHA to the version.